Update and rename new_deviation_function.m to calculation_deviation_values.m
This function outputs the deviation values by calculating the deviation of every single patient from the reference population. The construction of the reference model was based on the reference group consisting of cohort 1 patients who had 6 months of survival. The deviation values (Δ) were calculated as the perpendicular distance from individual patient data to regression line of the reference model. The perpendicular distance was calculated based on the residuals of the regression line from individual patient data using the following formula Δ= (x×y)/√(x^2+y^2 ) (x and y are the residuals of regression model). 

1. import datasets
2. deletion of missing values in rows for the pairwise comparison of variables
3. construction of regression model using the reference population
4. calculation of deviation values
5. output columns of deviation values aligning to the patient ID